Fields generated by MetaAnalytic now exclude codes with children. Fixed some issues when selecting the parent and dimension of a new analytic code. Added a new 'translate' option.
When creating a code, the parent cannot be selected (readonly) if the code's dimension is not selected (code view) or not a record in the database (dimension view, while creating a new dimension).

Added descriptions to models. Added a 'translate' option to enable/disable translation of analytic code and dimension names (default: False). Fields generated by the metaclasses now have a common placeholder string value. Cleaned up the translation file.