Analytic structure: added the get_structures and get_dimensions utility methods.
get_dimensions returns a dictionary with the dimension ID as key and their ordering as value. It can be useful in order to find the field associated with a specific dimension (obtained by external ID or by another model's get_dimension)

get_structures simply returns the structure browse records directly.