Update analytic_structure to v8
- Few small easy mistake fix (typo)
- Define _bound_dimension_id in _setup_complete method instead of __init__

The method calls get_object_reference or _update try to look up information
in the _columns of the object. But at '__init__' time, the object is not properly
finish and _columns is None.
By moving last initialisation in _setup_complete we are sure that object is initialized.

We need to keep the old api to be sure that 'self' is really the object.
If v8 api 'self' is a records set and we can't set _bound_dimension_id.

--HG--
branch : odoo8